% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LatentSpace.R
\docType{package}
\name{LatentSpace-package}
\alias{LatentSpace}
\alias{LatentSpace-package}
\title{LatentSpace: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
